EIEIO refactor
I wanted to try something new, and having read so much about CLOS, I
wanted to see how good it was working with it in elisp.
The result? Well, I added some more functionality with just a few more
lines of code and I think it is now easier to add some new backend (as
Windows) in a future.
All in all, one of the best thing this refactor brought was the
cl-function `counsel-spotify-search-query`. I think it exposes a good
API for a searching function with different options (IMHO)