Make asynchronous calls to Spotify
This commit brings asynchronous calls to the Spotify API. Hopefully typing in
the minibufer will feel snapier after this change.
I had to upgrade the dependency on Ivy because in the version that was used the
callback to update the Ivy candidates didn't exist yet.

Another thing I noticed doing this refactor was that for every call to Spotify,
the plugin negotiates its auth token. I recall doing it this way so i wouldn't
have to manage token expiration times and all that stuff. I think that is the
next step, if I want to make the plugin a little bit faster.