Reify authorization code flow and client credential flow
The central piece of this implementation is the
`counsel-spotify-request` generic function that has two
specializations: one for `client-credentals-flow` and one for
`authorization-code-flow`. Each API call can be called with one of this two
authorization flows in order to make a request to the Spotify servers in an
authorized way

TODO: Request can only be a GET request (because that is the only one we were
using up until noe). We need to support a couple of HTTP verbs to be able to
talk properly with Spotify APIs

Foo